backend/requirements.txt
# TruthLens Backend — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Production server (Gunicorn manages uvicorn workers) ──────────────────────
gunicorn==22.0.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# ── Auth & security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Config & validation ───────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
email-validator

# ── HTTP client + OpenAI (Stage 2 — graceful fallback if key missing) ─────────
httpx==0.27.2
openai==1.51.0

# ── Testing (excluded from Docker image — see Dockerfile) ─────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
